# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Build8Now Logistics, Loyalty & RBAC API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API for Build8Now material procurement. Features data-driven shipping profiles, multi-criteria freight calculation, 3-tier influencer loyalty precedence, database idempotency, append-only ledger, and role-based / object-level authorization (RBAC/ABAC)."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/influencers": {
      "get": {
        "summary": "List all registered influencer partners (Admin Only)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Influencers (Task 2)"],
        "responses": {
          "200": { "description": "List of influencer partners" },
          "403": { "description": "Forbidden - Admin role required" }
        }
      }
    },
...
    "/orders": {
      "get": {
        "summary": "List purchase orders (Customer views own orders; Admin views all orders)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Orders"],
        "responses": {
          "200": { "description": "List of orders" }
        }
      },
      "post": {
        "summary": "Place a new material purchase order with automated freight derivation",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items", "shippingAddress", "distanceKm"],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": { "type": "string", "example": "prod-cement-1" },
                        "quantity": { "type": "integer", "example": 20 }
                      }
                    }
                  },
                  "shippingAddress": { "type": "string", "example": "Site #42, DLF Phase 5, Gurugram" },
                  "distanceKm": { "type": "number", "example": 35.0 },
                  "influencerReferralCode": { "type": "string", "example": "INF-RAHUL-MAIN" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order placed with calculated shipping and linked influencer" }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "List available construction material products with assigned shipping profiles",
        "tags": ["Products Catalog"],
        "responses": {
          "200": { "description": "Array of products" }
        }
      },
      "post": {
        "summary": "Create a new product in the catalog (Admin Only)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Products Catalog"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slug", "sku", "category", "price", "weightKg"],
                "properties": {
                  "name": { "type": "string", "example": "Tata Tiscon 550D TMT Rebar 12mm" },
                  "slug": { "type": "string", "example": "tata-tiscon-550d-12mm" },
                  "sku": { "type": "string", "example": "STEEL-TATA-12MM" },
                  "category": { "type": "string", "example": "Steel & Structural" },
                  "price": { "type": "number", "example": 72.0 },
                  "weightKg": { "type": "number", "example": 10.65 },
                  "shippingProfileId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Product created successfully" }
        }
      }
    },
...
  }
}